Line Visualization Complete & Node Deletion
Finished line visualization. Nodes now have a line that appears and accurately (to my knowledge) displays the connections between nodes. Works with updating and moving nodes.

When moving a node the line updates to the dragged position.

Deletion of nodes is implemented. It deletes the node, updating said nodes connections and line visualizer.

There is also a isGraphComplete check. It goes through node and sees if a nodes slot is not connected. If so it returns false.

-- Some more checking is needed to see about left over objects / memory leaks after deletion.