Reworked NodeLineConnectionManager
NodeLineConnectionManager now draws to one single canvas.

The use of the manager has also been moved. Instead of each node having a line manager now the circuit manager class creates a line manager.

The line manager now loops over each node and draws the line. The only issue is it will draw a line twice, this will be fixed before push to main.

The node output slot now emit signals that are connected to the circuit manager for updating lines.

Plan to make it so the input slot can be used to connect nodes, instead of only the input slot.

Also small change, moved the circuit elements menu position on creation. So its not pos (0,0) / over the toolbar.