Added Sandbox Level & Other Small Changes
Added the sandbox level. Its the current iteration of a level used for testing. This will stay as the sandbox level. Other levels you won't be able to create an output node.

Added level names to the window title.

Added draggable window padding to the top of the window. So a node doesn't lay over the title bar. (Will need to be changed for MacOS, sense title bar is not in the window.)