Created Box2DManager
Box2DManager is the class that is responsible for the Box2D physics of a level.

Currently nodes can interact with one another when the physics is enabled. The simulation runs off a timer created in the Box2DManager class.

Currently the Box2D is only setup for the Sandbox level.