Node Quick Connect / Create
Created the base of the node quick connect / create system.

This system works via dragging an output into an empty space. A small ui appears with buttons of nodes to connect to. Clicking one creates the node and connects it to the node that was being dragged.

Some more work is needed but this is basically complete, need to implement this for input slots as well.

Also need a means to making the ui disappear when mouse cursor leaves the ui.