Box2D Scaling Fix & Other Changes
Fixed bug where scaling the nodes while physics was on would result in the box2d bodies not scaling as well. Now on scale the body fixtures are changed to scale as needed.

Also added a feature where on starting  the physics the state of the nodes (position and scale) are stored. So once the physics is stopped the nodes go back into place.

Fixed line placement on scaled nodes, when starting a drag from a node slot.

Fixed Box2D node body not being updated properly when a node is created from a quick connect.