DragableWidget Border Changes &  Zoom Fix
Made the border detection for dragable widget use the size of the widget. Making it so that when scaling it can be moved to the edges of the screen and no further. Before it would have a large gap when scaled.

Added box2d scale call for the zoom widget.